Shiv patch (#4)
Added documentation for distribution with shiv

Inlcudes minor, general changes and rearrangements.
In addition to the issues discussed in the pull request I moved the configuration section to the very top. Users now usually dont have to scroll through sections that are unimportant for them. Since most users will only need to read the "configuration", "use as zip application" and "usage" sections, it seems easier to have the easiest setup up top.